Make IdentifyIterators use IterationPrimaryClassInfo instead of hardcoding. (There's still a fair amount of hardcoding elsewhere!)
Unfortunately, we really want to just be listing interfaces and
classes in the XML and catching anything that implements that
interface or extends that class as well.  Currently JModeller is not
recognizing that the ReverseFigureEnumerator instances implement
FigureEnumeration.  So the next step will be to get a class hierarchy
into memory, either using reflection or dumping at instrumentation
time (preferably the latter).